make hero component wrap different children component, modification of css for responsive design
