+++ src/Components/Button/Button.css
@@ -4,6 +4,7 @@
   font: inherit;
   color: white;
   padding: 0.75em 3em;
+  border: 1px solid white;
   border-radius: 100vw;
   background-color: var(--pastelLilac);
   cursor: pointer;
