+++ src/Pages/Home.jsx
@@ -1,6 +1,11 @@
 import React from "react";
 import { Hero } from "../Components/Hero/Hero";
+import { Base } from "../Components/Banner/Base";
 
 export const Home = () => {
-  return <Hero />;
+  return (
+    <Hero>
+      <Base />
+    </Hero>
+  );
 };
